Draft a more general foreign function interface
The idea is for the foreign function state (and compiler) to be
parameterised on the function signature of FFI calls, rather than
requiring every foreign function to have the exact same signature.
The intent is to get rid of all the unverified marshalling/unmarshalling
code in basis_ffi.c and, hopefully, allow the FFI calls to be linked
directly to stdlib, with the compiler generating appropriate
marshalling/unmarshalling code based on the signatures.

This commit updates the source semantics with a first draft of what
this semantics might look like. This should be cleaned up, written in
lem, and updated with more types later.

The type signatures are intented to be intermediate between CakeML
types and C types, and only include types that map in a natural and
obvious way to a type in both languages.